+++ tests/WorkspaceFileResolver.test.cpp
@@ -1,0 +1,44 @@
+#include "doctest.h"
+#include "LSP/WorkspaceFileResolver.hpp"
+#include "Luau/Ast.h"
+#include "Luau/FileResolver.h"
+
+TEST_SUITE_BEGIN("WorkspaceFileResolverTests");
+
+TEST_CASE("resolveModule handles LocalPlayer PlayerScripts")
+{
+    WorkspaceFileResolver fileResolver;
+
+    Luau::ModuleInfo baseContext{"game/Players/LocalPlayer/PlayerScripts"};
+    auto resolved = fileResolver.resolveModule(&baseContext,
+        &Luau::AstExprIndexName(Luau::Location(), nullptr, Luau::AstName("PurchaseClient"), Luau::Location(), Luau::Position(0, 0), '.'));
+
+    REQUIRE(resolved.has_value());
+    CHECK_EQ(resolved->name, "game/StarterPlayer/StarterPlayerScripts/PurchaseClient");
+}
+
+TEST_CASE("resolveModule handles LocalPlayer PlayerGui")
+{
+    WorkspaceFileResolver fileResolver;
+
+    Luau::ModuleInfo baseContext{"game/Players/LocalPlayer/PlayerGui"};
+    auto resolved = fileResolver.resolveModule(
+        &baseContext, &Luau::AstExprIndexName(Luau::Location(), nullptr, Luau::AstName("GuiScript"), Luau::Location(), Luau::Position(0, 0), '.'));
+
+    REQUIRE(resolved.has_value());
+    CHECK_EQ(resolved->name, "game/StarterGui/GuiScript");
+}
+
+TEST_CASE("resolveModule handles LocalPlayer StarterGear")
+{
+    WorkspaceFileResolver fileResolver;
+
+    Luau::ModuleInfo baseContext{"game/Players/LocalPlayer/StarterGear"};
+    auto resolved = fileResolver.resolveModule(
+        &baseContext, &Luau::AstExprIndexName(Luau::Location(), nullptr, Luau::AstName("GearScript"), Luau::Location(), Luau::Position(0, 0), '.'));
+
+    REQUIRE(resolved.has_value());
+    CHECK_EQ(resolved->name, "game/StarterPack/GearScript");
+}
+
+TEST_SUITE_END();